+++ packages/components/src/utils/select/keyboard-behavior.spec.ts
@@ -9,7 +9,6 @@ import {
   getUpdatedIndex,
   getUsableSelectOptions,
   setHighlightedSelectOption,
-  setMatchingSelectOptionHighlighted,
   setNextSelectOptionHighlighted,
 } from './keyboard-behavior';
 
@@ -47,17 +46,21 @@ const generateOptions = (
         highlighted: highlightedIndex === idx,
         disabled: disabledIndex === idx,
         hidden: hiddenIndex === idx,
+        style: {
+          display: 'block',
+        },
       }) as unknown as Option
   );
 };
-
-// handleDropdownScroll is not tested on purpose
-// describe('handleDropdownScroll()', () => {});
 
 // getActionFromKey is not tested on purpose
 // describe('getActionFromKey()', () => {});
 
 describe('getUpdatedIndex()', () => {
+  it('should return -1 if maxIndex is -1', () => {
+    const index = getUpdatedIndex(1, -1, 'First');
+    expect(index).toBe(-1);
+  });
   it('should return 0 if action is First', () => {
     const index = getUpdatedIndex(1, 10, 'First');
     expect(index).toBe(0);
@@ -97,9 +100,10 @@ describe('getUpdatedIndex()', () => {
 });
 
 describe('setNextSelectOptionHighlighted()', () => {
-  it('should call correct functions to set the next option highlighted', () => {
+  it('should highlight the new option if no option is currently highlighted', () => {
     const options = generateOptions();
-    const listElement = document.createElement('ul');
+
+    options[1].scrollIntoView = jest.fn();
 
     const getHighlightedSelectOptionIndexSpy = jest
       .spyOn(keyboardBehaviorUtils, 'getHighlightedSelectOptionIndex')
@@ -108,21 +112,21 @@ describe('setNextSelectOptionHighlighted()', () => {
       .spyOn(keyboardBehaviorUtils, 'getUsableSelectOptions')
       .mockReturnValueOnce(options);
     const setHighlightedSelectOptionSpy = jest.spyOn(keyboardBehaviorUtils, 'setHighlightedSelectOption');
-    const handleSelectDropdownScroll = jest.spyOn(keyboardBehaviorUtils, 'handleSelectDropdownScroll');
 
     options.forEach((option) => {
       expect(option.highlighted).toBeFalsy();
     });
 
-    setNextSelectOptionHighlighted(listElement, options, 1);
+    setNextSelectOptionHighlighted(options, 1);
     expect(getHighlightedSelectOptionIndexSpy).toHaveBeenCalledWith(options);
     expect(getUsableSelectOptionsSpy).toHaveBeenCalledWith(options);
     expect(setHighlightedSelectOptionSpy).toHaveBeenCalledWith(options[1], true);
-    expect(handleSelectDropdownScroll).toHaveBeenCalledWith(listElement, options[1]);
+    expect(options[1].scrollIntoView).toHaveBeenCalledWith({ block: 'nearest' });
   });
-  it('should call setHighlightedSelectOption() twice when highlighted option exists', () => {
+  it('should update highlight by removing the highlight from the current option and highlighting the new one', () => {
     const options = generateOptions();
-    const listElement = document.createElement('ul');
+
+    options[1].scrollIntoView = jest.fn();
 
     const getHighlightedSelectOptionIndexSpy = jest
       .spyOn(keyboardBehaviorUtils, 'getHighlightedSelectOptionIndex')
@@ -131,18 +135,38 @@ describe('setNextSelectOptionHighlighted()', () => {
       .spyOn(keyboardBehaviorUtils, 'getUsableSelectOptions')
       .mockReturnValueOnce(options);
     const setHighlightedSelectOptionSpy = jest.spyOn(keyboardBehaviorUtils, 'setHighlightedSelectOption');
-    const handleSelectDropdownScroll = jest.spyOn(keyboardBehaviorUtils, 'handleSelectDropdownScroll');
 
     options.forEach((option) => {
       expect(option.highlighted).toBeFalsy();
     });
 
-    setNextSelectOptionHighlighted(listElement, options, 1);
+    setNextSelectOptionHighlighted(options, 1);
     expect(getHighlightedSelectOptionIndexSpy).toHaveBeenCalledWith(options);
     expect(getUsableSelectOptionsSpy).toHaveBeenCalledWith(options);
     expect(setHighlightedSelectOptionSpy).toHaveBeenCalledWith(options[1], true);
     expect(setHighlightedSelectOptionSpy).toHaveBeenCalledWith(options[2], false);
-    expect(handleSelectDropdownScroll).toHaveBeenCalledWith(listElement, options[1]);
+    expect(options[1].scrollIntoView).toHaveBeenCalledWith({ block: 'nearest' });
+  });
+  it('should remove highlight from the current option if no new option is to be highlighted', () => {
+    const options = generateOptions();
+
+    const getHighlightedSelectOptionIndexSpy = jest
+      .spyOn(keyboardBehaviorUtils, 'getHighlightedSelectOptionIndex')
+      .mockReturnValueOnce(2);
+    const getUsableSelectOptionsSpy = jest
+      .spyOn(keyboardBehaviorUtils, 'getUsableSelectOptions')
+      .mockReturnValueOnce(options);
+    const setHighlightedSelectOptionSpy = jest.spyOn(keyboardBehaviorUtils, 'setHighlightedSelectOption');
+
+    options.forEach((option) => {
+      expect(option.highlighted).toBeFalsy();
+    });
+
+    setNextSelectOptionHighlighted(options, -1);
+    expect(getHighlightedSelectOptionIndexSpy).toHaveBeenCalledWith(options);
+    expect(getUsableSelectOptionsSpy).toHaveBeenCalledWith(options);
+    expect(setHighlightedSelectOptionSpy).not.toHaveBeenCalledWith(options[1], true);
+    expect(setHighlightedSelectOptionSpy).toHaveBeenCalledWith(options[2], false);
   });
 });
 
@@ -213,31 +237,6 @@ describe('getMatchingSelectOptionIndex()', () => {
   });
 });
 
-describe('setMatchingSelectOptionHighlighted()', () => {
-  it('should call setNextSelectOptionHighlighted() when matching index found', () => {
-    const options = generateOptions();
-    const listElement = document.createElement('ul');
-    const getMatchingSelectOptionIndexSpy = jest
-      .spyOn(keyboardBehaviorUtils, 'getMatchingSelectOptionIndex')
-      .mockReturnValueOnce(2);
-    const setNextSelectOptionHighlightedSpy = jest.spyOn(keyboardBehaviorUtils, 'setNextSelectOptionHighlighted');
-    setMatchingSelectOptionHighlighted(listElement, options, 'a');
-    expect(getMatchingSelectOptionIndexSpy).toHaveBeenCalledWith(options, 'a');
-    expect(setNextSelectOptionHighlightedSpy).toHaveBeenCalledWith(listElement, options, 2);
-  });
-
-  it('should not call setNextSelectOptionHighlighted() when matching index not found', () => {
-    const options = generateOptions();
-    const listElement = document.createElement('ul');
-    const getMatchingSelectOptionIndexSpy = jest
-      .spyOn(keyboardBehaviorUtils, 'getMatchingSelectOptionIndex')
-      .mockReturnValueOnce(-1);
-    const setNextSelectOptionHighlightedSpy = jest.spyOn(keyboardBehaviorUtils, 'setNextSelectOptionHighlighted');
-    setMatchingSelectOptionHighlighted(listElement, options, 'a');
-    expect(getMatchingSelectOptionIndexSpy).toHaveBeenCalledWith(options, 'a');
-    expect(setNextSelectOptionHighlightedSpy).not.toHaveBeenCalled();
-  });
-});
 describe('setHighlightedSelectOption()', () => {
   it('should set option highlighted=true and call forceUpdate', () => {
     const forceUpdateSpy = jest.spyOn(stencilUtils, 'forceUpdate');